Delete .github/PULL_REQUEST_TEMPLATE.md (#5)
no more empty stuff like 

## Context/Change

Please include a summary of the change along with the reasoning and/or
motivation behind it. List any dependencies that are required and any
potential issues to look out for.

## Checklist

- [ ] 🧪 Ensure you have written tests
- [ ] 🧑‍🔬 Test your code in the appropriate environment (i.e
development/staging/production)
- [ ] 🗨️ Inform PM when PR is merged if the change is relevant to the
rest of the company

## Reference

Linear will automatically add a reference to the issue if you use the
name of the ticket as the branch.